PR Feedback: Consolidate credentials init in ITUtil.
